Deleted old text_extractor script and renamed new so that it takes its place
